The rest of the implementation
