+++ src/Neptuo.Productivity.SnippetManager.UI/InlineSnippetProvider.cs
@@ -6,7 +6,6 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows.Forms;
 
 namespace Neptuo.Productivity.SnippetManager;
