+++ src/Neptuo.Productivity.SnippetManager/CompositeSnippetProvider.cs
